Fix the Rubocop version of generated daemons
Since the newer versions of Rubocop have the Ruby on Rails-related cops
exported from the main gem and into a separate gem, generating a new
daemon doesn't make its Rubocop settings working out of the box.

This is fixed by specifying a minimal version of the main Rubocop gem
and including the 'rubocop-rails' gem in the generator's Gemfile.